+++ qa-tests-backend/src/test/groovy/ImageScanningTest.groovy
@@ -125,6 +125,9 @@ class ImageScanningTest extends BaseSpecification {
     ]
 
     def setupSpec() {
+        if (!shouldSpecRun(this.class.getSimpleName())) {
+            return
+        }
         ImageIntegrationService.deleteStackRoxScannerIntegrationIfExists()
         removeGCRImagePullSecret()
         ImageIntegrationService.deleteAutoRegisteredGCRIntegrationIfExists()
@@ -149,6 +152,9 @@ class ImageScanningTest extends BaseSpecification {
     }
 
     def cleanupSpec() {
+        if (!shouldSpecRun(this.class.getSimpleName())) {
+            return
+        }
         orchestrator.deleteNamespace(TEST_NAMESPACE)
 
         ImageIntegrationService.addStackroxScannerIntegration()
